DAT/requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
python-nmap>=0.7.1
Jinja2>=3.1.0
